Change * 2 on random fxn in getComputerChoice
The * 2 only allowed for two options rather than the 3 needed, simply changing it to 3 instead should fix the problem. NOTE: Pretty sure this is due to the fact that we are basically multiplying that number by the returned value that's less than 1, greater than 0. If we floor that, any time we multiply 2 by a number less than 0 it will return 1, leaving us with just 0 and 1 as our options. That's why increasing it to 3 allows for the proper distribution.